YARN-4844. Add getMemorySize/getVirtualCoresSize to o.a.h.y.api.records.Resource. (wangda)

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-common/src/main/java/org/apache/hadoop/yarn/util/resource/ResourceCalculator.java
@@ -31,18 +31,18 @@ public abstract class ResourceCalculator {
   public abstract int 
   compare(Resource clusterResource, Resource lhs, Resource rhs);
   
-  public static int divideAndCeil(int a, int b) {
+  public static long divideAndCeil(long a, long b) {
     if (b == 0) {
       return 0;
     }
     return (a + (b - 1)) / b;
   }
 
-  public static int roundUp(int a, int b) {
+  public static long roundUp(long a, long b) {
     return divideAndCeil(a, b) * b;
   }
 
-  public static int roundDown(int a, int b) {
+  public static long roundDown(long a, long b) {
     return (a / b) * b;
   }
 
@@ -54,7 +54,7 @@ public abstract class ResourceCalculator {
    * @param required required resources
    * @return number of containers which can be allocated
    */
-  public abstract int computeAvailableContainers(
+  public abstract long computeAvailableContainers(
       Resource available, Resource required);
 
   /**
@@ -169,7 +169,7 @@ public abstract class ResourceCalculator {
    * @param denominator denominator
    * @return resultant resource
    */
-  public abstract Resource divideAndCeil(Resource numerator, int denominator);
+  public abstract Resource divideAndCeil(Resource numerator, long denominator);
   
   /**
    * Check if a smaller resource can be contained by bigger resource.
